Add support for async tasks functional tests.
Includes initial support for geopackage export functional tests.

+++ src/web/functional/src/test/java/org/geogig/web/functional/FunctionalTestContext.java
@@ -10,7 +10,10 @@
 package org.geogig.web.functional;
 
 import java.io.File;
+import java.io.IOException;
 import java.net.URI;
+import java.util.HashMap;
+import java.util.Map;
 
 import org.junit.rules.ExternalResource;
 import org.junit.rules.TemporaryFolder;
@@ -30,6 +33,7 @@ import org.restlet.data.Method;
 import org.restlet.data.Reference;
 import org.restlet.data.Request;
 import org.restlet.data.Response;
+import org.w3c.dom.Document;
 
 import com.google.common.base.Preconditions;
 import com.google.common.base.Throwables;
@@ -47,27 +51,34 @@ public class FunctionalTestContext extends ExternalResource {
 
     private Response lastResponse;
 
-    @Override
-    public void before() throws Exception {
-        GlobalContextBuilder.builder = new CLIContextBuilder();
-        this.tempFolder = new TemporaryFolder();
-        this.tempFolder.create();
+    private Map<String, String> variables = new HashMap<>();
 
-        File rootFolder = tempFolder.getRoot();
-        repoProvider = new DirectoryRepositoryProvider(rootFolder);
-        this.app = new Main(repoProvider, true);
-        this.app.start();
+    @Override
+    public synchronized void before() throws Exception {
+        if (app == null) {
+            GlobalContextBuilder.builder = new CLIContextBuilder();
+            this.tempFolder = new TemporaryFolder();
+            this.tempFolder.create();
+
+            File rootFolder = tempFolder.getRoot();
+            repoProvider = new DirectoryRepositoryProvider(rootFolder);
+            this.app = new Main(repoProvider, true);
+            this.app.start();
+        }
     }
 
     @Override
-    public void after() {
-        try {
-            // this.client.stop();
-            this.app.stop();
-        } catch (Exception e) {
-            throw Throwables.propagate(e);
-        } finally {
-            tempFolder.delete();
+    public synchronized void after() {
+        if (app != null) {
+            try {
+                // this.client.stop();
+                this.app.stop();
+            } catch (Exception e) {
+                throw Throwables.propagate(e);
+            } finally {
+                this.app = null;
+                tempFolder.delete();
+            }
         }
     }
 
@@ -96,7 +107,19 @@ public class FunctionalTestContext extends ExternalResource {
         return testData;
     }
 
-    public void call(final Method method, final String resourceUri) {
+    public void call(final Method method, String resourceUri) {
+
+        this.lastResponse = callInternal(method, resourceUri);
+    }
+
+    public Response callDontSaveResponse(final Method method, String resourceUri) {
+        return callInternal(method, resourceUri);
+    }
+
+    private Response callInternal(final Method method, String resourceUri) {
+
+        resourceUri = replaceVariables(resourceUri, this.variables);
+
         Request request = new Request(method, resourceUri);
         request.setRootRef(new Reference(""));
         if (Method.PUT.equals(method) || Method.POST.equals(method)) {
@@ -104,11 +127,72 @@ public class FunctionalTestContext extends ExternalResource {
             // CommandResource at all
             request.setEntity("empty payload", MediaType.TEXT_PLAIN);
         }
-        this.lastResponse = app.handle(request);
+        Response response = app.handle(request);
+        return response;
+    }
+
+    public void setVariable(String name, String value) {
+        this.variables.put(name, value);
+    }
+
+    public String getVariable(String name) {
+        return getVariable(name, this.variables);
+    }
+
+    static public String getVariable(String varName, Map<String, String> variables) {
+        String varValue = variables.get(varName);
+        Preconditions.checkState(varValue != null, "Variable " + varName + " does not exist");
+        return varValue;
+    }
+
+    public String replaceVariables(final String text) {
+        return replaceVariables(text, this.variables);
+    }
+
+    static String replaceVariables(final String text, Map<String, String> variables) {
+        String resource = text;
+        int varIndex = -1;
+        while ((varIndex = resource.indexOf("{@")) > -1) {
+            for (int i = varIndex + 1; i < resource.length(); i++) {
+                char c = resource.charAt(i);
+                if (c == '}') {
+                    String varName = resource.substring(varIndex + 1, i);
+                    String varValue = getVariable(varName, variables);
+                    String tmp = resource.replace("{" + varName + "}", varValue);
+                    resource = tmp;
+                    break;
+                }
+            }
+        }
+        return resource;
     }
 
     public Response getLastResponse() {
         Preconditions.checkState(lastResponse != null);
         return lastResponse;
     }
+
+    public String getLastResponseText() {
+        String xml;
+        try {
+            xml = getLastResponse().getEntity().getText();
+        } catch (IOException e) {
+            throw Throwables.propagate(e);
+        }
+        return xml;
+    }
+
+    public String getLastResponseContentType() {
+        final String xml = getLastResponse().getEntity().getMediaType().getName();
+        return xml;
+    }
+
+    public Document getLastResponseAsDom() {
+        try {
+            return getLastResponse().getEntityAsDom().getDocument();
+        } catch (IOException e) {
+            throw Throwables.propagate(e);
+        }
+    }
+
 }
